+++ packages/editor-cq/docs/CHANGELOG.md
@@ -1,6 +1,12 @@
 # @atlaskit/editor-cq
 
-## Unreleased
+## 4.5.1 (2017-08-03)
+
+* bug fix; fix browserstack tests import not transpiled editor-core ([3e729c2](https://bitbucket.org/atlassian/atlaskit/commits/3e729c2))
+
+
+
+
 
 ## 4.5.0 (2017-07-28)